chore: make createProfilesChangedListener more abstract
