+++ src/social-network-provider/options-page/index.ts
@@ -4,7 +4,7 @@ import Services from '../../extension/service'
 import { MessageCenter } from '../../utils/messages'
 import { GetContext, ValueRef } from '@holoflows/kit/es'
 import { Profile } from '../../database'
-import { createProfilesChangedListener } from '../../social-network/defaults/FriendsValueRef'
+import { createDataWithIdentifierChangedListener } from '../../social-network/defaults/createDataWithIdentifierChangedListener'
 
 const optionsPageUISelf = defineSocialNetworkUI({
     ...emptyDefinition,
@@ -23,7 +23,7 @@ const optionsPageUISelf = defineSocialNetworkUI({
             const ref = optionsPageUISelf.friendsRef
             MessageCenter.on(
                 'profilesChanged',
-                createProfilesChangedListener(ref, () => true),
+                createDataWithIdentifierChangedListener(ref, () => true),
             )
             Services.Identity.queryProfiles().then(p => (ref.value = p))
         }
